Initial Commit
Big thanks to the following for this initial commit:

https://github.com/alicoding/react-webpack-babel

The react, webpack, babel starter kit was very quick, easy, straight
forward and gave me enough to start from scratch.

https://github.com/freddyrangel/playing-with-react-and-d3

The playing with react and d3 starter kit had great steps to get you
familiar with the basics. Combined, these two are great to start learning
both react and d3. This was a little outdated and used d3 v3 and there
were a few breaking changes introduced in v4 that aren't addressed in this
repo.